neovim: don't show completion menu automatically
Update configuration for nvim-cmp so that the auto-completion menu
doesn't appear automatically, and change the mappings to confirm the
first suggested completion so that they don't interfere with indents
and returns while typing. Made this change because buffer completions
in particular are distracting when they pop up for almost every word
in long text documents.